+++ CHANGES.md
@@ -6,6 +6,8 @@ _tesla-microservice_ is used for a number of different services now. Still it is
 This release cleans up remnants of past eperiments and unused functionality. This leads to breaking changes. 
 
 - Remove ```register-timed-handler``` from ```handler``` namespace. Use ```goo/timing-middleware``` instead.
+- Remove ```register-response-fn``` from handler namespace. This was only used internally.
+- Move internally used middlewares to separate namespaces.
 
 
 ### 0.11.0