+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 _tesla-microservice_ is used for a number of different services now. Still it is a work in progress. This section will document changes and give instructions on breaking ones. Likely you will find corresponding changes in [tesla-examples](https://github.com/otto-de/tesla-examples).
 
+### 0.15.0
+Changed securing of internal endpoints. Moved from providing separate auth-functions
+to components app-status and metering to an auth-middleware provided to the base-system 
+see [Securing internal info endpoints](https://github.com/otto-de/tesla-microservice#securing-internal-info-endpoints)
+
 ### 0.14.0
 This release cleans up remnants of past eperiments and unused functionality. This leads to breaking changes.